add responsive pages

+++ frontend-foodstore/src/pages/address/index.jsx
@@ -10,6 +10,7 @@ import { NewAddress } from "./NewAddress";
 import { PrimaryButton } from "../../components/button/PrimaryButton";
 import PrimaryActionButton from "../../components/button/PrimaryActionButton";
 import SecondaryActionButton from "../../components/button/SecondaryActionButton";
+import Swal from "sweetalert2";
 
 export const Address = () => {
   const [address, setAddress] = useState([])
@@ -40,15 +41,38 @@ export const Address = () => {
 
   }
 
-  const handleDeleteAddress = async (addressId) => {
-    try {
-        const res = await addressService.deleteAddress(addressId)
-        if (res.status === 200) {
+  const handleDeleteAddress = async (addressId, e) => {
+    e.preventDefault()
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result)=>{
+      if(result.isConfirmed){
+        try {
+            await addressService.deleteAddress(addressId)
             setAddress(address.filter((address) => address._id !== addressId));
+            Swal.fire({
+              title: "Deleted!",
+              text: "Your address has been deleted.",
+              icon: "success",
+            }).then(() => {
+                window.location.reload();
+            });
+        } catch (err) {
+          console.error(err);
+          Swal.fire({
+              title: "Error!",
+              text: "An error occurred while deleting the address.",
+              icon: "error",
+          });
         }
-    } catch (error) {
-        console("Error deleting address:", error);
-    }
+      }
+    })
   };
 
 const handleEditAddress = (addressId, event) => {
@@ -73,6 +97,13 @@ const handleEditAddress = (addressId, event) => {
         );
         setShowEditModal(false);
         setAddressToEdit(null);
+        Swal.fire({
+            title: "Success!",
+            text: "Address Edited!",
+            icon: "success"
+        }).then(() => {
+            window.location.reload(); // Reload the page after the user clicks "OK"
+        });
       }
     } catch (error) {
       console.error("Error updating address:", error);
@@ -83,6 +114,13 @@ const handleEditAddress = (addressId, event) => {
     try {
       await addressService.storeAddress(newAddress)
       setShowNewModal(false);
+      Swal.fire({
+        title: "Success!",
+        text: "Add new address!",
+        icon: "success"
+    }).then(() => {
+        window.location.reload(); // Reload the page after the user clicks "OK"
+    });
     } catch (error) {
       console.error("Error adding address:", error);
     }
@@ -92,7 +130,7 @@ const handleEditAddress = (addressId, event) => {
   return (
     <div className="flex flex-row sm:flex-row">
     <SideBar/>
-    <div className="flex justify-center pt-10  md:pt-20 lg:pt-24 h-screen mx-auto">
+    <div className="flex justify-center pt-10  md:pt-20 lg:pt-24 h-screen md:mx-auto ml-24 mr-4">
       <div className="w-full" >
         <h1 className="text-center pb-3 font-poppins font-bold text-3xl md:text-4xl lg:text-5xl">Address</h1>
         <div className="border-1 border-[#000] p-2 flex gap-1 rounded-md">
@@ -104,20 +142,21 @@ const handleEditAddress = (addressId, event) => {
                 <div>
                     <table className="w-full border-collapse">
                         <thead className=" border-b-2">
-                            <tr>
+                            <tr className="md:text-[18px] text-[15px]">
                                 <th className="p-2">Nama</th>
                                 <th className="p-2">Detail</th>
                             </tr>
                         </thead>
                                 <tbody>
                           {address.map((address) => (
-                            <tr key={address._id}>
+                            <tr key={address._id} className="md:text-[15px] text-[12px]">
                               <td className="p-2">{editAddressId === address._id ? <input type="text" className="text-black" value={address.nama} onChange={(e) => setAddress(address.map(a => a._id === address._id ? { ...a, nama: e.target.value } : a))} /> : address.nama}</td>
                               <td className="p-2">{editAddressId === address._id ? <input type="text" className="text-black" value={address.detail} onChange={(e) => setAddress(address.map(a => a._id === address._id ? { ...a, detail: e.target.value } : a))} /> : `${address.detail} ${address.kelurahan}, ${address.kecamatan}, ${address.kabupaten}, ${address.provinsi}`}</td>
                               <td className="p-2">
-                                <PrimaryActionButton onClick={(e) => handleEditAddress(address._id, e)}>Edit</PrimaryActionButton>
-                                <span> </span>
-                                <SecondaryActionButton onClick={() => handleDeleteAddress(address._id)}>Delete</SecondaryActionButton>
+                                <div className="flex gap-2">
+                                  <PrimaryActionButton onClick={(e) => handleEditAddress(address._id, e)}>Edit</PrimaryActionButton>
+                                  <SecondaryActionButton onClick={(e) => handleDeleteAddress(address._id, e)}>Delete</SecondaryActionButton>
+                                </div>
                               </td>
                             </tr>
                           ))}
